src/requirements.txt
uvicorn
fastapi
typing
pillow
torch
torchvision
python-multipart
opencv-python-headless